backend/requirements.txt
fastapi[standard]
uvicorn
passlib
python-jose[cryptography]
python-multipart
google-generativeai
pydantic
bcrypt==3.2.0
python-dotenv
pytrends
PyJWT